refactor: sanitize wizards
