+++ apps/extension/src/ui/domains/Staking/Bittensor/BittensorChangeLockHotkeyModal/BittensorChangeLockHotkeyConfirm.tsx
@@ -1,5 +1,3 @@
-import type { SignerPayloadJSON } from "@core/domains/signing/types"
-import type { DotNetworkId } from "@talismn/chaindata-provider"
 import { AlertCircleIcon } from "@talismn/icons"
 import { ScrollContainer } from "@ui/components/ScrollContainer"
 import { WizardModalDialog } from "@ui/components/WizardModalDialog"
@@ -12,33 +10,11 @@ import { SapiSendButton } from "@ui/domains/Transactions/SapiSendButton"
 import { cn } from "@ui/util/cn"
 import type { FC } from "react"
 import { useTranslation } from "react-i18next"
-import type { Hex } from "viem"
 
 import {
   BITTENSOR_CHANGE_LOCK_HOTKEY_MODAL_CONTAINER_ID,
-  type ConvictionMoveOutcome,
-} from "./BittensorChangeLockHotkeyContent"
-
-type BittensorChangeLockHotkeyConfirmProps = {
-  networkId: DotNetworkId
-  address: string
-  currentHotkey: string
-  destinationHotkey: string
-  /** the resolved conviction consequence of the move, surfaced as the confirm-step notice */
-  convictionOutcome: ConvictionMoveOutcome
-  lockedAmount: bigint
-  symbol: string
-  tokenId: string
-  taoTokenId: string | null | undefined
-  payload: SignerPayloadJSON | undefined
-  txMetadata: Uint8Array | `0x${string}` | undefined
-  feeEstimate?: bigint | null
-  isLoadingFeeEstimate?: boolean
-  errorFeeEstimate?: unknown
-  onBack: () => void
-  onClose: () => void
-  onSubmitted: (hash: Hex) => void
-}
+  useBittensorChangeLockHotkeyWizard,
+} from "./useBittensorChangeLockHotkeyWizard"
 
 const HotkeyDisplay: FC<{ hotkey: string }> = ({ hotkey }) => (
   <span className="inline-flex max-w-full items-center gap-4 overflow-hidden align-middle">
@@ -54,26 +30,35 @@ const SummaryRow: FC<{ label: string; children: React.ReactNode }> = ({ label, c
   </div>
 )
 
-export const BittensorChangeLockHotkeyConfirm: FC<BittensorChangeLockHotkeyConfirmProps> = ({
-  networkId,
-  address,
-  currentHotkey,
-  destinationHotkey,
-  convictionOutcome,
-  lockedAmount,
-  symbol,
-  tokenId,
-  taoTokenId,
-  payload,
-  txMetadata,
-  feeEstimate,
-  isLoadingFeeEstimate,
-  errorFeeEstimate,
-  onBack,
-  onClose,
-  onSubmitted,
-}) => {
+export const BittensorChangeLockHotkeyConfirm = () => {
   const { t } = useTranslation()
+  const {
+    networkId,
+    address,
+    existingLock,
+    selectedHotkey,
+    convictionOutcome,
+    symbol,
+    baseTokenId,
+    taoTokenId,
+    payload,
+    txMetadata,
+    feeEstimate,
+    isLoadingFeeEstimate,
+    errorFeeEstimate,
+    close,
+    setStep,
+    onSubmitted,
+  } = useBittensorChangeLockHotkeyWizard()
+
+  if (
+    !address ||
+    !existingLock ||
+    !selectedHotkey ||
+    !convictionOutcome ||
+    convictionOutcome === "checking"
+  )
+    return null
 
   // the conviction consequence is the headline of this step: a hard warning when it resets, a calm
   // note when it's preserved or boosted. Either way, spell out that only the hotkey moves.
@@ -106,8 +91,8 @@ export const BittensorChangeLockHotkeyConfirm: FC<BittensorChangeLockHotkeyConfi
     <WizardModalDialog
       title={t("Conviction Lock Hotkey")}
       contentClassName="size-full flex flex-col overflow-hidden"
-      onBackClick={onBack}
-      onCloseClick={onClose}
+      onBackClick={() => setStep("form")}
+      onCloseClick={close}
     >
       <ScrollContainer className="grow" innerClassName="flex w-full flex-col gap-8">
         <h2 className="mb-4 text-center font-bold text-md">{t("Review transaction")}</h2>
@@ -117,15 +102,15 @@ export const BittensorChangeLockHotkeyConfirm: FC<BittensorChangeLockHotkeyConfi
             <StakingAccountDisplay address={address} chainId={networkId} className="text-sm" />
           </SummaryRow>
           <SummaryRow label={t("Current hotkey")}>
-            <HotkeyDisplay hotkey={currentHotkey} />
+            <HotkeyDisplay hotkey={existingLock.hotkey} />
           </SummaryRow>
           <SummaryRow label={t("New hotkey")}>
-            <HotkeyDisplay hotkey={destinationHotkey} />
+            <HotkeyDisplay hotkey={selectedHotkey} />
           </SummaryRow>
           <SummaryRow label={t("Locked amount")}>
             <TokensAndFiat
-              planck={lockedAmount}
-              tokenId={tokenId}
+              planck={existingLock.amount}
+              tokenId={baseTokenId}
               noCountUp
               tokensClassName="text-body"
             />
