+++ apps/extension/src/ui/domains/Staking/Bittensor/BittensorChangeLockTypeModal/index.tsx
@@ -1,12 +1,34 @@
 import { Modal } from "@ui/components/Modal"
 import { PopupSizeModalContainer } from "@ui/components/PopupSizeModalContainer"
-import type { FC } from "react"
+import { SuspenseTracker } from "@ui/components/SuspenseTracker"
+import { TxProgress } from "@ui/domains/Transactions/TxProgress"
+import { type FC, Suspense } from "react"
 
 import { useBittensorChangeLockTypeModal } from "../hooks/useBittensorChangeLockTypeModal"
+import { BittensorChangeLockTypeConfirm } from "./BittensorChangeLockTypeConfirm"
+import { BittensorChangeLockTypeForm } from "./BittensorChangeLockTypeForm"
 import {
   BITTENSOR_CHANGE_LOCK_TYPE_MODAL_CONTAINER_ID,
-  BittensorChangeLockTypeContent,
-} from "./BittensorChangeLockTypeContent"
+  BittensorChangeLockTypeWizardProvider,
+  useBittensorChangeLockTypeWizard,
+} from "./useBittensorChangeLockTypeWizard"
+
+const BittensorChangeLockTypeRouter = () => {
+  const { step, hash, networkId, close } = useBittensorChangeLockTypeWizard()
+
+  switch (step) {
+    case "form":
+      return <BittensorChangeLockTypeForm />
+    case "confirm":
+      return <BittensorChangeLockTypeConfirm />
+    case "submitted":
+      return hash ? (
+        <div className="size-full p-12">
+          <TxProgress hash={hash} networkIdOrHash={networkId} onClose={close} />
+        </div>
+      ) : null
+  }
+}
 
 export const BittensorChangeLockTypeModal: FC = () => {
   const { isOpen, args, close } = useBittensorChangeLockTypeModal()
@@ -15,13 +37,14 @@ export const BittensorChangeLockTypeModal: FC = () => {
     <Modal isOpen={isOpen && !!args} onDismiss={close}>
       <PopupSizeModalContainer id={BITTENSOR_CHANGE_LOCK_TYPE_MODAL_CONTAINER_ID}>
         {!!args && (
-          <BittensorChangeLockTypeContent
+          <BittensorChangeLockTypeWizardProvider
+            // reset wizard state whenever it is opened for a different subnet/account
             key={`${args.networkId}-${args.netuid}-${args.address ?? ""}`}
-            networkId={args.networkId}
-            netuid={args.netuid}
-            seedAddress={args.address}
-            onClose={close}
-          />
+          >
+            <Suspense fallback={<SuspenseTracker name="BittensorChangeLockTypeModal" />}>
+              <BittensorChangeLockTypeRouter />
+            </Suspense>
+          </BittensorChangeLockTypeWizardProvider>
         )}
       </PopupSizeModalContainer>
     </Modal>
