+++ apps/extension/src/ui/domains/Staking/Bittensor/components/BittensorLockTypePicker.tsx
