+++ apps/extension/src/ui/domains/Staking/Bittensor/components/ConvictionLockHotkeyPicker.tsx
