+++ apps/extension/src/ui/domains/Staking/Bittensor/components/HotkeyPickerRows.tsx
